% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHAlogrithmMulti.R
\name{MHAlogrithmmulti}
\alias{MHAlogrithmmulti}
\title{Metropolis-Hastings Algorithm to sample data from multivariate distribution}
\usage{
MHAlogrithmmulti(
  TargetDensity,
  nvar,
  CGDensity = "multivariateNormal",
  xinit = NULL,
  sigma = NULL,
  niter = 1000
)
}
\arguments{
\item{TargetDensity}{The PDF of target density function from which the users want to sample. This should be as a function of x.}

\item{nvar}{The number of variables}

\item{CGDensity}{Candidate Generating Density function (default is multivariate normal distribution)}

\item{xinit}{The starting point of the algorithm}

\item{sigma}{The covariance matrix of the candidate generating distribution}

\item{niter}{The number of data to sample}
}
\value{
Return the data collection sampled from the MH Algorithm
}
\description{
Metropolis-Hastings Algorithm to sample data from multivariate distribution
}
