% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHAlogrithmUni.R
\name{MHAlogrithmuni}
\alias{MHAlogrithmuni}
\title{Metropolis-Hastings Algorithm to sample data from univariate distribution}
\usage{
MHAlogrithmuni(
  TargetDensity,
  CGDensity = "Normal",
  xinit = 0,
  sigma = 1,
  niter = 1000
)
}
\arguments{
\item{TargetDensity}{The PDF of target density function from which the users want to sample. This should be as a function of x.}

\item{CGDensity}{Candidate Generating Density function (default is normal distribution)}

\item{xinit}{The starting point of the algorithm}

\item{sigma}{The standard deviation of the Candidate Generating Density}

\item{niter}{The number of data to sample}
}
\value{
Return the data collection sampled from the MH Algorithm
}
\description{
Metropolis-Hastings Algorithm to sample data from univariate distribution
}
